requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
python-multipart==0.0.6
torch>=1.8.0
torchvision>=0.9.0
numpy>=1.20.0
Pillow>=9.0.0
matplotlib>=3.4.0
ultralytics>=8.0.0
lime>=0.2.0
scikit-image>=0.19.0
opencv-python>=4.5.0
python-jose[cryptography]>=3.3.0
aiofiles>=0.8.0
pytest>=7.0.0
requests>=2.28.0
uuid>=1.30
starlette>=0.27.0


# cd backend
# uvicorn app:app --reload

# cd frontend
# python -m http.server 8080